Statically link LuaXML_lib into FlyWithLua plugin binary.
This is the safest way to ensure that the binary module is both compiled
and linked against exactly the same version of LuaJIT as used by
FlyWithLua.

The conventional method [1] of building and linking binary Lua modules
does not work well in case of FlyWithLua. Normally the host executable
would provide a Lua implementation and any binary Lua module is supposed
to link against symbols exported from the host at runtime. This is
impossible in case of X-Plane and FlyWithLua because FWL is a
dynamically loaded binary itself. Moreover, it is most likely to be not
the only plugin with Lua code (e.g. aircrafts using XLua or SASL), so we
must not export Lua symbols into public space to avoid messy conflicts.

All and all, it is the safest for us to link whatever binary Lua modules
we want to support and resort onto dynamically loaded binaries only in a
worst case (e.g. users adding their own modules).

All those considerations, however, don't impact normal, non-binary Lua
modules, since linking is no concern for them.

[1]: http://lua-users.org/wiki/BuildingModules