Support passing callbacks to floating window functions by reference.
This allows using local or even anonymous functions as callbacks, which
is convenient in complicated scope and prevents global namespace
pollution.

This commit uses sol2 to support both old and new style of passing
callbacks as well as for invoking Lua callbacks safely. One of the
things which happens behind the scenes is that sol::protected_function
keeps the function from being garbage-collected even if it is not
referenced anywhere else (by using Lua Registry feature). Doing so in
plain would Lua C API have been pretty tedious, but with sol2 it comes
for free.

Overall this commit serves as a decent showcase for sol2 benefits:
 1) Ability to bind C++ functions with arbitrary signatures to Lua,
    handling Lua stack transparently.
 2) Automatic function type checking when calling C++ from Lua.
 3) Argument type-based overloading: picking right C++ implementation
    based on argument types passed from Lua.
 4) Safely calling Lua functions from C++ with the same syntax as normal
    C++ functions and handling errors.