Basic framework for consistent error handling.
Functions in the runtime namespace provide a robust way to catch and
handle errors triggered by Lua scripts. In future all call to Lua will
go through this API, rather than low-level C API.

The new API can unwind Lua call stack and make a reasonable guess for
which script is responsible for triggering an error and quarantine it
automatically.

For now, it is mostly in proof-of-concept state, handling errors at
script loading stage and in do_often() code. Some functionality is
missing and reporting errors to user is done through Dev Console only.

I also did a bunch of refactoring to the code I had to change, hopefully
making it easier to understand and use more modern C++.