Create a GitHub Actions workflow for building and packaging FWL.
The new workflow replaces previous Docker-based cross-compilation build
system and should simplify FWL development and pull request evaluation.

C++ cross-compilation is often challenging, especially for project with
external library dependencies, especially when building for Mac on
Linux. The new workflow builds FWL for Mac and Windows natively on the
corresponding platform using corresponding GH Actions runners, which
solves a lot of issues with toolchain setup.

CMake scripts were adjusted such that the standard `cmake && cmake
--build` works on each supported platform without any additional setup.
This should be sufficient for local development and testing, and pushing
the code to GitHub will produce a complete FluWithLua_plugin.zip, which
can be used for cross-platform testing.

Some notable changes in this commit:

  - Compiled *.xpl files are no longer included in the repository. They
    used to be included for the convenience of manual packaging process,
    but are no longer needed since packages are build automatically on
    push.
  - If a new release is created on GitHub, the workflow will build and
    attach a plugin package to the release automatically. This should
    make releasing new FWL versions somewhat less work.
  - OS X SDK version is now selected via `CMAKE_OSX_DEPLOYMENT_TARGET`
    variable, which is set at the top of `src/CMakeLists.txt`. In future
    we will be able to select 11.3 version when building for M1 Macs.

One interesting future improvement opportunity this opens up is building
FWL using MSVC on Windows, which is the complier recommended by XPlane
developers. This may (or may not) produce a better-performing code for
Windows.